# teamflow-ui/src/app/services/models/project-request.ts
/* tslint:disable */
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

export interface ProjectRequest {
  description?: string;
  teamEmails?: Array<string>;
  title: string;
}