# teamflow-ui/src/app/services/models/project.ts
/* tslint:disable */
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

import { User } from '../models/user';
export interface Project {
  createdAt?: string;
  description?: string;
  id?: number;
  owner?: User;
  teamMembers?: Array<User>;
  title?: string;
  updatedAt?: string;
}