# teamflow-ui/src/app/services/models/role.ts
/* tslint:disable */
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

export interface Role {
  createdDate?: string;
  id?: number;
  lastModifiedData?: string;
  name?: string;
}