# teamflow-ui/src/app/services/models/task.ts
/* tslint:disable */
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

import { User } from '../models/user';
export interface Task {
  assignees?: Array<User>;
  createdAt?: string;
  deadline?: string;
  description?: string;
  id?: number;
  status?: 'TODO' | 'IN_PROGRESS' | 'DONE' | 'BLOCKED';
  title?: string;
  updatedAt?: string;
}